Simplify statements internal objects
We pull the list of plans and spaces from the billing events without
using the CF API. There is no need to generate objects with
`metadata` and `entity` subobjects when we only want a list of
`guid` and `name` of spaces and plans.

We simplify these objects and update the view and tests accordingly.

We also change the logic to verify plan uniqueness based on GUID,
instead of name.

We also need to change the test data in billing.test.data.ts to
have 2 events with the same plan and ensure 100% coverage.

Signed-off-by: henrytk <henry.knott@digital.cabinet-office.gov.uk>