Copy init.js script to assets folder
When we moved the GDS frontend init call to a separated file, we
referred to a file in src. but that file is not copied when the
app is built and pussed to CF :(

Instead, we must copy the file as part of webpack. We remove the
middleware route for `src` as it is not needed.

Fixes #100

See 64f9a098086d47b316552d7f729cd17c331d887e