Move the location of the middleware
We have found that the location of this middleware matters. From the
documentation we're gathering that the middleware needs to be run after
the initial body parser is kicking in. We're accidentally doing that
when converting the urlencoded froms to extended version.

Alternatively, this should have been done just before each handler being
executed, which means we'd need to modify our router. - We'd rather not
do that at this time.