Update from alpha to beta
I think the paas-admin phase banner was set to `ALPHA` because
paas-admin was a new service that we're getting feedback on.

From the point of view of a tenant this is confusing though, because
they don't see paas-admin as being a separate thing to the PaaS. This
means it looks like the whole PaaS is in alpha, which it isn't. For
example:

http://govuk.zendesk.com/agent/tickets/3411752

Furthermore, paas-admin itself really feels like it's in beta now - it's
available to all our tenants, new tenants are using it by default, and
we're not planning on throwing the implementation away to build a new
one.