Tighten types for invitation emails
In #111 we fixed a bug where we hadn't passed `location` into the
welcome email personalisation. This could have been caught by the type
system, but we were using the loose `IParameters` type (which just says
"an object with some keys") and the even looser `any` type (which can be
assigned to anything).

I had to make the UAAClient declare the type of the inviteUser response,
otherwise the `inviteLink` parameter would have had the `any` type, but
we know it has to be a string.

I didn't bother checking what all the fields on the UAAClient inviteUser
response were - I've only typed the two fields we actually use.