Render header with plain HTML
We want to be able to pass it a location tag into the header, but the
function from the GOV.UK Frontend module is very strict in what it
takes, so we swap it out for HTML from the same official source[1].

The actual location tag will be added in a future commit. This is
therefore effectively a no-op.

[1] https://design-system.service.gov.uk/components/header/