Add location to govuk template
We want to add a location tag (Ireland or London) to the main header of
every page. Because every page extends this template we have to pass the
value in to every page rendering. For the error handling middleware this
is less easy, because they do not have the app context, which contains
the configuration information. That is why we grab the AWS region
information from the environment variable and duplicate the mapping of
that to a label (this ideally she only be done once in `src/main`.