Determine platform location in main
We want to use the AWS region to show which platform is being used in
the header. This means the human-readable labels (Ireland and London)
will need to be in the app context, as it will be passed to every
template. Therefore, we move the translation of AWS region into a
location name to main, where it only needs to be done once.