Fix bug for users with org roles but no space roles
Because we were populating our object by iterating over spaces, we'd
miss any users that had org-level permissions, but weren't assigned to
any of the spaces the logged in user had access to.

This is easy to fix by populating the object by iterating over users
instead, we just need a map of spaces by user.

To make the code more readable I've extracted a `groupByAndMap` function
into `lib`. I think it's unlikely this code is useful enough that it
will be used again, but inlining the function at the call sight makes
the code even more unreadable than it already is.