Add tests and fix code to match intended behaviour
The "functional" code before was only functional in the computer-science
sense - it didn't actually function.

I've added some more tests to cover the more complicated use cases and
switched to a more imperative style, which in this case I think improves
the readability and fixes the correctness issues we were having before.

In particular the code now handles the case where a user belongs to more
than one space, which it didn't before.