Resolve disabled field = intent to remove perm
As we disable the checkbox if the user is the last billing manager / org manager (so the last remaining user can't be removed), the value of this field is not serialised.

This meant that the last billing user will *always* have their permission removed if their permissions page is submitted.

This is resolved by adding a hidden field with the same name and value of the checkbox if the checkbox is disabled - therefore the field value is still posted on submission.