Get Quota from the billing response
The paas-admin cost report currently queries the cf api for the quota guid and the
billing endpoint for the events.

Paas-billing has been updated
(https://github.com/alphagov/paas-billing/pull/64) to include the
quota_definition_guid so it has the correct quota for the billing period. This
updates the cost report to pull the quota from the billing response if
available.