Enable strict null checks, fix possible issue
Previously we disabled the strictNullChecks option because it was
warning us about passing a `(string | undefined)` into a function that
wanted a `(string)`.

It looks like the compiler was warning us about a genuine issue here
though:

* If there were any billable events for a particular org in the time
  period, `orgBillableEvent` would be defined, and `orgBillableEvent[0]`
  would be defined
* The event could be one from before we started writing `quotaGUID` to
  the databases, so `orgBillableEvent[0].quotaGUID` could still be
  undefined
* The conditional would have set `orgQuotaGUID` to undefined in this
  case.
* Calling `cf.organizationQuota` with undefined would result in an
  error.

This commit pulls the code to get the first billable event out into a
function so it's a bit clearer what it's doing. This function is clear
about the fact that it can return undefined. We can then use the this
thing as the first option in the conditional, and fall back to the
current quota if that doesn't work.