Add S3 Buckets to the pricing calculator
We are introducing a new backing service which provides AWS S3 buckets
to our tenants. We have configured an initial pricing plan [1] which
is working in Cloud Foundry and in `paas-billing`.

This commit makes those S3 bucket pricing plans visible in our Pricing
Calculator. This is useful because it will allow (prospective) tenants
to understand how much our S3 buckets will cost them.

The new row in the pricing calculator will be entitled
`aws-s3-bucket default`, in line with the pricing plan [1]. This is
comparable to existing names such as `postgres 10.5` [2].

[1] https://github.com/alphagov/paas-cf/commit/9cafbaeeba36798b7763a0f9f7b7f3ee9962e021

[2] https://admin.cloud.service.gov.uk/calculator