Don't specify default plan versions in Calculator
Our S3 bucket service only has one pricing plan, named
`aws-s3-bucket default`. It is visually strange to be specifying
`default` in the Pricing Calculator if there is *only* that default.
This commit suppresses printing `default` when a service has only one
pricing plan and it's version is named `default`.