Implement skeleton stub billing API
To work out whether a user is an admin we use `cf.hasOrganizationRole`,
which looks up the user by GUID.

By setting the user to have the same GUID that the stub organisation api
returns we can make paas admin think they're an admin, and then show
them the billing tab.

In this commit we're just adding the bear-minimum of the billing API to
stop it from erroring.