Refactor Calculator's use of pricing plan names
The pricing plans used by the Calculator come from the Billing API,
not Cloud Controller. They have a `name` field and the Calculator
deconstructs that to get various metadata about them.

The implementation was scattered and undocumented. This PR refactors
how this metadata extraction is done and documents how it works (see
the large comment.)

This commit exposes complexity that already existed. It should make
further refactoring easier, make the Calculator more maintainable,
and trigger better ways of storing metadata in `paas-billing`.