Remove currency rate and fix code to GBP
This commit updates `paas-admin` for changes to `paas-billing`.

The Billing API has been returning prices in GBP, but with a
`currency_code` and `currency_rate` field suggesting they were in GBP.
We are removing the `currency_rate` field from the Billing API in
favour of a `/currency_rates` API. We are also correcting the
`currency_code` field to say GBP. [1]

In this commit, all test data for `/billing_events` is altered to
remove `currency_rate` and set `currency_rate` to GBP. This commit
also starts getting currency rates from the new `/currency_rates`
endpoint, and VAT rates from the new `/vat_rates`.

[1] https://github.com/alphagov/paas-billing/pull/66