Uses `optimize-cssnano-plugin` to minify CSS
When upgrading `css-loader` from 1.0.0 to 2.1.1, `css-loader` began throwing an
error when discovering additional (unexected) properties in its `options`
parameter.

In particular, it was erroring on the prescence of the `minimize` option that
was removed in 1.0.0, but which we were supplying as `NODE_ENV ===
'production'` to enable it in production. As per the guidance in
`css-loader` v1.0.0 release notes [1], the `optimize-cssnano-plugin` is now
used to achieve CSS minification in production builds.

[1] https://github.com/webpack-contrib/css-loader/releases/tag/v1.0.0