Ignore errors when trying to transpile `.test.ts` files
At some point between versions 4.4.2 and 5.4.4, `ts-loader` began reporting
errors for `.test.ts` files. These errors surrounded not being able to find
names for jest and Jasmine matchers, and are resolved by the test environment
which is present when the tests are run. Thus, it is safe to ignore these
errors when trying to transpile non-test TS.