Warn about cflinuxfs2 on org, space and app pages
We need applications to stop using the `cflinuxfs2` stack. It has
reached end-of-life because it is based upon Ubuntu Trusty. This
commit adds highly-visible warnings to our admin panel when tenant
apps are still on `cflinuxfs2`.

The warnings are shown on organisation pages, space pages and on the
application pages themselves.

The homepage listing organisations does not display these warnings.
That would require persisting some state or for that page to request
every application the user has permissions to see. Neither would be
quick and low-risk to implement.

This commit hardcodes the GUID of `cflinuxfs2` in our `prod` and
`prod-lon` environments in order to keep things simple. This commit
will be reverted once the migration to `cflinuxfs3` is complete.