Add stacks (e.g., cflinuxfs3) endpoints to CF lib
Our Cloud Controller API library should be able to get information on
Stacks. Application endpoints only return the GUID of the stack they
are using, and this change will allow us to provide the friendly name
of the stack. This will be useful in encouraging tenants to upgrade to
`cflinuxfs3`.