Remove vulnerability scanning from test runs
In Travis, vulnerability scanning was being performed as part of a test run,
but didn't appear to alert us to anything. However, locally and on Concourse
we're being alerted and blocked by what it finds.

Since we've never had that information from Travis, and we now have Dependobot
installed for exactly this purpose, I've removed the vulnerability scanning
from test runs. I've left the NPM script in place for future developers.