Allow the filtering of cost reports by service
This allows you to go to `/reports/cost/2019-03-01?service=blah` and see
only the bills for that particular service.

The particular use case for this is to see how much we've billed for
elasticsearch.

It works in a pretty hacky way, using the query param as a substring to
search through all the plan names.