Link region name to the multi-region sign-in page
It is helpful that `paas-admin` now says "IRELAND" or "LONDON" in the
top-right. But users did not have a readily available way to change
environment.

This commit makes the "IRELAND" or "LONDON" text a link leading to the
page on our product page which gives users a choice of which
environment to sign into [1].

[1] https://www.cloud.service.gov.uk/sign-in