Run stub APIs on different ports
Expanding which endpoints we cover in our stubbing of the different APIs means
we've run in to conflicts, where two distinct APIs have the same path component
and different responses. This commit spins up the different stub APIs on
different ports so that those conflicts don't happen.

It also extracts setting the required environment variables in to a script, so
as not to pollute package.json.