Make session available on IContext
Session access will be required for the OIDC flow. It makes sense to make it
available to HTTP handler implementations via the passed-in IContext
implementation, because the session is part of the context for handling a
request. This commit adds the field, and refactors how tests make their
IContext implementations.