Remove dependency on `npm`
This was added in 13a30a74, but we don't actually need to have it in our dev
dependencies. `npm` still needs to be installed globally if you're going
to do anything with it.

This removes a transitive dependency on a vulnerable version of fstream,
which brings `npm audit`s count of vulnerabilities down to 0.

This is also nice because npm is a thing that gets new versions a lot,
so it causes quite a bit of dependabot churn for no good reason.