Add report page showing cost breakdown by service
It's often useful to know how much individual organisations are spending
on groups of services (e.g. postgres or elasticsearch). This information
makes it easier to plan capacity and prioritise work on features that
are getting a lot of usage.

At the moment it's pretty painful to see this information in one place,
so this commit adds a new report page which will show the cost breakdown
by organisation and service for a given month.

It renders something like this:

```
Billables by service for July 2019
----------------------------------

| Service          | Including VAT | Excluding VAT |
|------------------|---------------|---------------|
| some-other-thing | £110000.00    | £0.00         |
| postgres         | £1000.00      | £0.00         |
| compute          | £111.00       | £0.00         |

Billables by organisation and service for July 2019
---------------------------------------------------

| Organization               | Service          | Including VAT | Excluding VAT |
|----------------------------|------------------|---------------|---------------|
| the-system_domain-org-name | some-other-thing | £10000.00     | £0.00         |
| the-system_domain-org-name | postgres         | £1000.00      | £0.00         |
| the-system_domain-org-name | compute          | £111.00       | £0.00         |
| unknown                    | some-other-thing | £100000.00    | £0.00         |
```

This new report should have roughly the same performance as the existing
`/reports/cost/` endpoint - possibly a bit faster as it doesn't need to
get the quotas.