Sankey diagram visualisation of billing flow
I couldn't get webpack to bundle D3 nicely with the client side sankey
code, so we're just loading it in a script tag from node_modules
instead.

client side sankey code is deliberately written to be ES5 compatible,
although it's questionable whether this is necessary since the audience
is GDS people who'll be using latest Firefox / Chrome.

I've reused `getBillableEventsByOrganisationAndService` from
`cost-by-service.ts` - it would probably be nice to do some
restructuring of the files rather than reaching into a different
"component" for this function but I can't be bothered for now.