Fix incorrect numberOfNodes for HA services
Some services (at the moment just redis medium-ha-3.2) have multiple
nodes, and the price depends on the number of nodes.

Currently the calculator serializes the "state" to a set of hidden
inputs, which are submitted along with the form when the user adds an
app or service.

Unfortunately, the `numberOfNodes` submitted for new services is always
`plans[0].numberOfNodes` - i.e. the number of nodes for the smallest
plan, which is invariably 1.

When the user is choosing a redis plan they only interact with one
select, where they pick the plan guid, so there's no way of them also
submitting the number of nodes.

Fortunately we don't actually need the user to submit how many nodes
there are for a particular service plan, because the correct value will
always still be available on the plan we got back from the API.

This commit continues to use the serialized numberOfNodes and
memoryInMB for apps, but stops sending them for services and uses
the value from the plans instead.