Prevent double update of username in MS SSO uplift journey
When performing a Microsoft SSO uplift journey, the user's username was being
updated twice because of a missing break statement. First, it was updated to
the value of the `oid` claim (correctly), it was then falling through and being
updated to the value of the `sub` claim (incorrectly).

Added a pair of expectations to the tests to verify that only one call is
happening, and refactored the code to remove the possibility of a second call
to `uaa.setUserOrigin`.

This _may_ have been happening for Google users too, but I haven't verified
that. If it was, the code path was leaving the account in the right state
because `google` was the second and last, case.