Add origin to Org / Users / Team members
As an org manager or org auditor of an organisation I want to be able to
see who in my organisation has enabled single sign on, so I can enforce
a security policy.

I chose to introduce `ul.plain` because the regular `govuk-list` class
looks very strange on this page, when we move to GOV.UK Frontend 3.0 we
can address this page a bit more.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>