Don't bother with builders
After a discussion with @richardtowers we can just use lodash.merge
everywhere, we aren't getting much by wrapping it in a class.

If we return a new object (rather than JSON string) with default values,
we can use it everywhere in our tests, whilst also getting the benefits
of the type system.

by default lodash.merge mutates the leftmost object so in order to
safely reuse the defaultOrg you have to lodash.merge({}, org, ...) which
is a footcannon

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>