Parse pricing plan formulae directly
This uses a [peg.js](https://pegjs.org/) grammar to generate a parser
for the formulae in the pricing plans (e.g. `ceil($time_in_seconds/3600) * 0.01`).

The parser does all of the maths when parsing the expression. A "parsed"
expression will evaluate to a single number.

This approach has a few major advantages:

* It's HUGELY faster than talking to billing
* We can remove all of the /forecast_events code from billing
* It reduces the load on the billing database
* We can test that the calculator is working properly within paas-admin
  (e.g. that it's getting the maths right)
* We get to do something that almost feels like computer science

There are some disadvantages though:

* There's a new opportunity for bugs now that we're not using the exact
  same expression evaluator as postgres (which could cause the
  calculator to give incorrect results, if we screw this up)
* If we add any logic to the pricing plans that isn't implemented in our
  simple grammar (e.g. if we used `floor` as well as `ceil`) then the
  calculator will break until we fix the grammar.

TODO: proper tests of the grammar