Make page layout more compact
We use quite a lot of vertical space unnecessarily and do not make full
use of our phase banner

Jeremy recommended I take a look at the Content Data service and see how
their navigation is laid out

Removed the "this is a new service" as the service has been around for
years.

Sending us feedback and getting support were the same link, so I've
combined them

Documentation is now called out separately not as part of the top level
navigation, as it exists on a different site.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>