Do not use headings in permissions details
We shouldn't use headings inside persissions details

The best way to do this would be to have a detail for each role, e.g.

> What is an org manager?
> What is an org billing manager?

etc

But I didn't do that because it is out of scope of just updating styles

This also gets rid of the weird 'heading-medium' 'heading-large' classes
we had knocking around unnecessarily.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>