Hide users who appear in CF, but not in UAA
There are scenarios where a user can be recorded as having permissions for an
organisation by CF, but not have a corresponding entry in UAA. Previously, this
would cause an error and a resultant error page. Now, we filter out CF users
who don't have UAA records, and log a warning when we encounter it.

We believe filtering out those users is the correct answer because the members
page is intended to show users who has access to their organisation. Without a
UAA record, the user can't log in to get access, so they are irrelevant in that
respect.

Fixes #377