Remove govuk-body from `<body>`
The .govuk-body class is only intended for body text (like paragraphs
and so on), not the `<body>` element. Putting it on the `<body>` adds a
bunch of unintended styles. See #426.

Because we've had the class on the body all along we've got used to the
default font-family being GDS Transport, and removing it now causes a
bunch of things to fall through to the browser's default (serif,
usually). Rather than going through and adding classes to all of these
elements it's easier just to set the font-family on the body.

`.govuk-body` did a bunch of other things which will subtly change the
layouts when we remove it. Better to show than tell, so I'll put some
screenshots in the PR for this commit.