Support multiple exchange rates per month
`paas-billing` is gaining support for changing the USD-GBP exchange
rate partway through a month. This means a monthly invoice may use
multiple exchange rates, and we need to display that.

This commit is an as-simple-as-possible way to render multiple
exchange rates. If there is only one exchange rate that month, nothing
will look any different. If there are multiple exchange rates they are
outputted along with their "valid from" times.

With one exchange rate:

> Exchange rate: £1 to $1.25

With multiple exchange rates:

> Exchange rate: £1 to $1.25 from January 1st 2017,
  £1 to $1.25 from January 15th 2017

This isn't pretty, but it will solve the need to disclose our exchange
rate for the time being.