Add backing service metrics for RDS services
Initially we're using GetMetricWidgetImage as a quick way of getting metrics in
front of users. We know these are not good for accessibility, so we'll replace
them with "proper" graphs in a future story.

We're only doing RDS services (postgres/mysql) to begin with because it's easy
to work out which AWS resource corresponds to a service in cloud foundry. This
is a bit more tricky for Redis because there's not an obvious mapping between
the service instance id and the elasticache cluster id.

We've decided not to test the image generation code because it's only temporary.