Check that the user can access service metrics
Without this check it would be possible for users to see each others'
services by fiddling with the URL to the image (if they knew the service
instance guid). Simplest way is just to make a call to the CF API to
check that they've really got access.

We chose not to add an automated test for this method because it's
temporary code, so testing was done manually in a dev env (open the
image in a new tab, then sign in as a user without access to the service
and reload the image).