Calculate cluster names with fnv-plus and base32
Because elasticache has a strict limit on the length of the cluster
name the broker uses a base32 encoded FNV-1a hash to transform the
service GUID to a name that AWS will accept.

Calculating these hashes in paas-admin means we don't need to give the
application any more AWS permission than it needs, and reduces the
number of AWS API calls we need to make.

The test data in this commit was generated using the elasticache-broker,
so we can be reasonably confident the implementations are the same.