Refactor building getMetricWidgetImage API input
We're not calling these functions in the controller yet, but adding the
implementation and tests.

Redis services can be High Availability (HA), in which case there will
be multiple clusters and therefore multiple metrics. Each of the
clusters will have a name starting with the `replicationGroupId` (e.g.
`cf-th3c2nitfdqoc`), but it may be appended with `-001`, `-002`, or
`-0001-001`. This means we have to use a `SEARCH` expression instead of
being able to pick out the exact metric we want.

If the Redis is HA then the SEARCH will return multiple time series,
which means the Redis metrics need to show legends (postgres/mysql
don't, because there's always a single series). The SEARCH also doesn't
return units, so we have to look those up for ourselves.

Since we're passing user input (the metric dimension will come from a
URL parameter) into our query it's a good idea to restrict the list of
possible values, so we've enumerated the `supportedDimensions` to
sanitize the input.