Add metrics for redis services
Adds a bunch of Redis metrics to the view, and refactors the code for
generating image tags a bit to make it less noisy.

Also updates `cf.test.data.ts` so the service label is "postgres" not
"postgres-58" so the example data returned by the stub api can be used
for testing. If you want to test redis with the stub api at the moment
you have to change this label in the test data locally.