Be restrictive when checking the boxes for user
At the moment, when no email is provided, all permissions are being
selected. This is somewhat annoying to our tenants and should not have
happened.

With the current change, the previous state will be applied upon error.

We're extending the test case to be assured that this is being monitored
in some shape or form.