Add audit event pages for space/app/service
Adds new pages for:

- /organisations/./spaces/./events/eventGUID
- /organisations/./spaces/./applications/./events/eventGUID
- /organisations/./spaces/./services/./events/eventGUID

so a user can get additional context about the event which has happened.

Just dumping the RAW JSON for the data is fine at this point.

A follow up story can link the events pages to the docs which show how
to get them via the API.

While writing this the common template which showed the events for
apps/spaces was no longer able to be common without injecting a param
which i thought was a bit gross, because now we link to the event
details. So I just inlined the template.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>