Converge onto single way of handling checkboxes
At the moment we're dealing with two ways of managing the permissions
via the Admin interface. This is confusing especially when it comes down
to Nunjucks and the lack of types...

Converging onto a single way of dealing with it in the typescript side,
should make it easier to follow for people.

We're fixing the bug with checkboxes being disabled, when there aren't
enough of managers in the organisation.

Closes #505

Writing some tests, to be assured that some bugs will not re-occur.