Add a resolver for metrics
We'd like to allow users to pick from a range of times latest to their
relative local time. For instance:

Last 1h, Last 3h, Last 24h, Last week, Last month

In order to assure this is indeed picking a range `NOW` - `OFFSET` at
all times, and not hack around with html/forms to inject ranges into the
system, we can have an additional endpoint that will be generating the
ranges for us and redirect back to the view with the set of ranges.

This for instance could have been a problem, if user has a stale view of
the page, say not refreshed for a day, and then they'd pick an offset
such as 15m. The pre-baked range would mean they're looking at 15m slot
one day ago.

This _should_ also solve the problem, of the server being too slow, and
the user manages to choose an offset and immediately attempt to update
the form with submit button.