Add the metrics ranges
We'd like to allow our users to choose specific range for their
representation of metrics, to help them better visualise what's going on
on our platform.

This is a rough design and should be revisited, hopefully with a
designer.

We're also reducing the size of the charts themselves, to perhaps clean
up a little visually and leave us with some room with potential
imporvements to the charts themselves, for instance a tooltip on a side.