Upgrade TypeScript
There is a problem with a new version of TypeScript that hasn't been
occuring before. It's related to Diagnostics being enabled and reporting
on some [practically non-existing issues][1], such as:

> Object is possibly 'undefined'

This could be avoided by adding the `!` symbol to the object which
tricks the `tsc` to trust the developer.

I see that as a scary practice, which could quickly turn into something
like `font-size: 10px !important;`...

Instead, we're telling `jest`, not to run these diagnostics itself, as
personally I trust that linter and `tsc` itself will be able to handle
it and these don't throw these errors.

Saying that, it makes me think there is some other configuration to
jest, that could allow us keep the diagnostics, as well as act the same
way the linter/tsc do.

Disabling it for now, seems like sane approach.

[1]: https://github.com/microsoft/TypeScript/issues/29642