Fix the labeling problem
When the labels for the xScale are being generated and injected into the
D3 model, there will be cases where D3 library is too smart for it's own
good. For instance, if the month has changed in the middle of the chart,
but the tick doesn't happen specifically on the beggining of the month,
D3 will decide, that this is important milestone to represent on the
chart.

We've tried different approaches, some of which include:

- Hiding the unwanted tick with css
- Hiding the unwanted tick with D3 api
- Changing the interval for the labels
- Changing the format of the tick

And settled on us rotating the labels, as the solution seems clean
enough and doesn't allow for problem's reoccurrance.

Now that we have a time ranges specification, d3 library can be used to
automatically decide what format the metrics should be presented in.

Co-authored-by: Rafal Proszowski <rafal.proszowski@digital.cabinet-office.gov.uk>
Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>